refactor: extract mutation handling into DefaultMutationHandler
Move handle_add/remove_processor and handle_add/remove_connection from
flow_engine.rs into DefaultMutationHandler in engine/mutation_handler.rs.
The run_mutation_handler loop now delegates to this struct, reducing
flow_engine.rs from 841 to ~555 lines. Apply cargo fmt to all changed
files.
