+++ crates/runifi-core/src/engine/flow_engine.rs
@@ -16,7 +16,8 @@ use super::handle::{
     RelationshipInfo,
 };
 use super::metrics::ProcessorMetrics;
-use super::mutation::{MutationCommand, MutationError};
+use super::mutation::MutationCommand;
+use super::mutation_handler::DefaultMutationHandler;
 use super::processor_node::{
     ProcessorNode, SchedulingStrategy, SharedInputConnections, SharedInputNotifiers,
     SharedOutputConnections,
@@ -464,7 +465,7 @@ impl FlowEngine {
 ///
 /// Keeps the concrete enum internal to the engine — callers receive a `String`
 /// so they do not need to import or match on `SchedulingStrategy`.
-fn scheduling_display(strategy: &SchedulingStrategy) -> String {
+pub(crate) fn scheduling_display(strategy: &SchedulingStrategy) -> String {
     match strategy {
         SchedulingStrategy::TimerDriven { interval_ms } => {
             format!("timer-driven ({}ms)", interval_ms)
@@ -475,6 +476,10 @@ fn scheduling_display(strategy: &SchedulingStrategy) -> String {
 
 // ── Mutation handler ──────────────────────────────────────────────────────────
 
+/// Drive the mutation command loop, delegating each command to `DefaultMutationHandler`.
+///
+/// Topology mutations are kept sequential (one command at a time) by this loop
+/// so that `DefaultMutationHandler` methods never run concurrently.
 #[allow(clippy::too_many_arguments)]
 async fn run_mutation_handler(
     mut rx: mpsc::Receiver<MutationCommand>,
@@ -488,7 +493,17 @@ async fn run_mutation_handler(
     parent_cancel: CancellationToken,
     proc_tokens: Arc<parking_lot::Mutex<HashMap<String, CancellationToken>>>,
 ) {
-    let mut runtime_conn_id: usize = 0;
+    let mut handler = DefaultMutationHandler {
+        live_procs,
+        live_conns,
+        content_repo,
+        id_gen,
+        bulletin_board,
+        registry,
+        parent_cancel,
+        proc_tokens,
+        runtime_conn_id: 0,
+    };
 
     loop {
         tokio::select! {
@@ -507,367 +522,35 @@ async fn run_mutation_handler(
                         name, type_name, properties,
                         scheduling_strategy, interval_ms, reply,
                     } => {
-                        let result = handle_add_processor(
+                        let result = handler.handle_add_processor(
                             &name, &type_name, properties,
                             &scheduling_strategy, interval_ms,
-                            &live_procs, &content_repo, &id_gen,
-                            &bulletin_board, &registry,
-                            &parent_cancel, &proc_tokens,
                         );
                         let _ = reply.send(result);
                     }
 
                     MutationCommand::RemoveProcessor { name, reply } => {
-                        let result = handle_remove_processor(
-                            &name, &live_procs, &live_conns, &proc_tokens,
-                        ).await;
+                        let result = handler.handle_remove_processor(&name).await;
                         let _ = reply.send(result);
                     }
 
                     MutationCommand::AddConnection {
                         source_name, relationship, dest_name, config, reply,
                     } => {
-                        runtime_conn_id += 1;
-                        let conn_id = format!("runtime-conn-{}", runtime_conn_id);
-                        let result = handle_add_connection(
-                            conn_id, &source_name, &relationship, &dest_name,
-                            config, &live_procs, &live_conns,
+                        handler.runtime_conn_id += 1;
+                        let conn_id = format!("runtime-conn-{}", handler.runtime_conn_id);
+                        let result = handler.handle_add_connection(
+                            conn_id, &source_name, &relationship, &dest_name, config,
                         );
                         let _ = reply.send(result);
                     }
 
                     MutationCommand::RemoveConnection { id, force, reply } => {
-                        let result = handle_remove_connection(&id, force, &live_conns);
+                        let result = handler.handle_remove_connection(&id, force);
                         let _ = reply.send(result);
                     }
                 }
             }
         }
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn handle_add_processor(
-    name: &str,
-    type_name: &str,
-    properties: HashMap<String, String>,
-    scheduling_strategy: &str,
-    interval_ms: u64,
-    live_procs: &Arc<RwLock<Vec<ProcessorInfo>>>,
-    content_repo: &Arc<dyn ContentRepository>,
-    id_gen: &Arc<IdGenerator>,
-    bulletin_board: &Arc<BulletinBoard>,
-    registry: &Option<Arc<PluginRegistry>>,
-    parent_cancel: &CancellationToken,
-    proc_tokens: &Arc<parking_lot::Mutex<HashMap<String, CancellationToken>>>,
-) -> std::result::Result<(), MutationError> {
-    // Fix 5: validate scheduling_strategy strictly.
-    if scheduling_strategy != "timer" && scheduling_strategy != "event" {
-        return Err(MutationError::InvalidSchedulingStrategy(
-            scheduling_strategy.to_string(),
-        ));
-    }
-
-    let reg = registry
-        .as_ref()
-        .ok_or_else(|| MutationError::Internal("No plugin registry available".into()))?;
-
-    let processor: Box<dyn Processor> = reg
-        .create_processor(type_name)
-        .ok_or_else(|| MutationError::UnknownType(type_name.to_string()))?;
-
-    if live_procs.read().iter().any(|p| p.name == name) {
-        return Err(MutationError::DuplicateName(name.to_string()));
-    }
-
-    let scheduling = if scheduling_strategy == "event" {
-        SchedulingStrategy::EventDriven
-    } else {
-        SchedulingStrategy::TimerDriven { interval_ms }
-    };
-
-    let metrics = Arc::new(ProcessorMetrics::new());
-    metrics
-        .enabled
-        .store(false, std::sync::atomic::Ordering::Relaxed);
-
-    let prop_descriptors: Vec<PropertyDescriptorInfo> = processor
-        .property_descriptors()
-        .into_iter()
-        .map(|pd| PropertyDescriptorInfo {
-            name: pd.name.to_string(),
-            description: pd.description.to_string(),
-            required: pd.required,
-            default_value: pd.default_value.map(|v| v.to_string()),
-            sensitive: pd.sensitive,
-            allowed_values: pd
-                .allowed_values
-                .map(|av| av.iter().map(|v| v.to_string()).collect()),
-        })
-        .collect();
-
-    let relationships: Vec<RelationshipInfo> = processor
-        .relationships()
-        .into_iter()
-        .map(|r| RelationshipInfo {
-            name: r.name.to_string(),
-            description: r.description.to_string(),
-            auto_terminated: r.auto_terminated,
-        })
-        .collect();
-
-    let shared_props = Arc::new(RwLock::new(properties));
-    let child_token = parent_cancel.child_token();
-
-    let pn = ProcessorNode::new(
-        name.to_string(),
-        format!("runtime-{}", name),
-        processor,
-        scheduling.clone(),
-        shared_props.clone(),
-        content_repo.clone(),
-        id_gen.clone(),
-        child_token.clone(),
-        metrics.clone(),
-        bulletin_board.clone(),
-    );
-
-    // Capture shared handles before moving pn into spawn.
-    let input_h = pn.input_connections_handle();
-    let output_h = pn.output_connections_handle();
-    let notifiers_h = pn.input_notifiers_handle();
-
-    proc_tokens.lock().insert(name.to_string(), child_token);
-    tokio::spawn(pn.run());
-
-    live_procs.write().push(ProcessorInfo {
-        name: name.to_string(),
-        type_name: type_name.to_string(),
-        scheduling_display: scheduling_display(&scheduling),
-        metrics,
-        property_descriptors: prop_descriptors,
-        relationships,
-        properties: shared_props,
-        input_connections: input_h,
-        output_connections: output_h,
-        input_notifiers: notifiers_h,
-    });
-
-    tracing::info!(name, type_name, "Hot-added processor");
-    Ok(())
-}
-
-/// Remove a processor at runtime.
-///
-/// Fix 3 (TOCTOU): Acquire the write lock for the entire operation so that
-/// concurrent lifecycle ops (start/stop) that check the same `live_procs` list
-/// cannot race with removal. This means a concurrent `start_processor` reading
-/// the processors list will either observe the processor as still present (before
-/// the write lock is acquired here) or not at all (after retain completes).
-async fn handle_remove_processor(
-    name: &str,
-    live_procs: &Arc<RwLock<Vec<ProcessorInfo>>>,
-    live_conns: &Arc<RwLock<Vec<ConnectionInfo>>>,
-    proc_tokens: &Arc<parking_lot::Mutex<HashMap<String, CancellationToken>>>,
-) -> std::result::Result<(), MutationError> {
-    // Check for active connections before acquiring the write lock.
-    // This avoids holding the write lock while iterating connections.
-    {
-        let conns = live_conns.read();
-        let conn_ids: Vec<String> = conns
-            .iter()
-            .filter(|c| c.source_name == name || c.dest_name == name)
-            .map(|c| c.id.clone())
-            .collect();
-        if !conn_ids.is_empty() {
-            return Err(MutationError::ProcessorHasConnections(conn_ids.join(", ")));
-        }
-    }
-
-    // Acquire the write lock for the remainder of the operation. This serialises
-    // removal with any concurrent lifecycle operation (start/stop/pause/resume)
-    // that holds a read lock while setting atomics.
-    let mut procs = live_procs.write();
-
-    let info = procs
-        .iter()
-        .find(|p| p.name == name)
-        .ok_or_else(|| MutationError::ProcessorNotFound(name.to_string()))?;
-
-    let enabled = info
-        .metrics
-        .enabled
-        .load(std::sync::atomic::Ordering::Relaxed);
-    let active = info
-        .metrics
-        .active
-        .load(std::sync::atomic::Ordering::Relaxed);
-    if enabled || active {
-        return Err(MutationError::ProcessorNotStopped);
-    }
-
-    if let Some(token) = proc_tokens.lock().remove(name) {
-        token.cancel();
-    }
-
-    procs.retain(|p| p.name != name);
-
-    tracing::info!(name, "Hot-removed processor");
-    Ok(())
-}
-
-/// Add a connection at runtime and wire it into the processor data paths.
-///
-/// Fix 1 (phantom connections): in addition to updating `live_conns`, push the
-/// new `FlowConnection` into:
-///   - the source processor's `output_connections` (SharedOutputConnections), and
-///   - the destination processor's `input_connections` (SharedInputConnections)
-///     and `input_notifiers` (SharedInputNotifiers).
-///
-/// These shared `Arc<RwLock<Vec<...>>>` are held inside the running
-/// `ProcessorNode` task, so writes here are immediately visible to the task's
-/// next invocation without restarting it.
-fn handle_add_connection(
-    conn_id: String,
-    source_name: &str,
-    relationship: &str,
-    dest_name: &str,
-    config: BackPressureConfig,
-    live_procs: &Arc<RwLock<Vec<ProcessorInfo>>>,
-    live_conns: &Arc<RwLock<Vec<ConnectionInfo>>>,
-) -> std::result::Result<String, MutationError> {
-    // Validate processors/relationship and collect the shared handles we'll
-    // mutate. Clone the Arcs so we can drop the read lock before writing.
-    let (src_output_h, src_rel, dst_input_h, dst_notifiers_h) = {
-        let procs = live_procs.read();
-
-        let src = procs
-            .iter()
-            .find(|p| p.name == source_name)
-            .ok_or_else(|| MutationError::ProcessorNotFound(source_name.to_string()))?;
-
-        // Find the matching RelationshipInfo to reconstruct a Relationship value.
-        let src_rel_info = src
-            .relationships
-            .iter()
-            .find(|r| r.name == relationship)
-            .ok_or_else(|| {
-                MutationError::UnknownRelationship(
-                    relationship.to_string(),
-                    source_name.to_string(),
-                )
-            })?;
-
-        let dst = procs
-            .iter()
-            .find(|p| p.name == dest_name)
-            .ok_or_else(|| MutationError::ProcessorNotFound(dest_name.to_string()))?;
-
-        // Relationship uses &'static str; we must leak the strings for runtime-
-        // allocated names. These strings live for the lifetime of the process,
-        // which is acceptable because connection/processor names are bounded by
-        // the configured max (128 chars) and the total number of hot-added
-        // connections is expected to be small.
-        use runifi_plugin_api::relationship::Relationship;
-        let rel = Relationship {
-            name: Box::leak(src_rel_info.name.clone().into_boxed_str()),
-            description: Box::leak(src_rel_info.description.clone().into_boxed_str()),
-            auto_terminated: src_rel_info.auto_terminated,
-        };
-
-        (
-            Arc::clone(&src.output_connections),
-            rel,
-            Arc::clone(&dst.input_connections),
-            Arc::clone(&dst.input_notifiers),
-        )
-    };
-
-    // Check for duplicate connection.
-    {
-        let conns = live_conns.read();
-        if conns.iter().any(|c| {
-            c.source_name == source_name
-                && c.relationship == relationship
-                && c.dest_name == dest_name
-        }) {
-            return Err(MutationError::DuplicateConnection(
-                source_name.to_string(),
-                relationship.to_string(),
-                dest_name.to_string(),
-            ));
-        }
-    }
-
-    let fc = Arc::new(FlowConnection::new(conn_id.clone(), config));
-
-    // Wire into destination processor's input list and notifier list.
-    // The running ProcessorNode task holds the same Arc<RwLock<...>>, so writes
-    // here are visible on its next `input_connections.read()` / `input_notifiers.read()`.
-    let notifier = fc.notifier();
-    dst_input_h.write().push(Arc::clone(&fc));
-    dst_notifiers_h.write().push(notifier);
-
-    // Wire into source processor's output list.
-    src_output_h.write().push((src_rel, Arc::clone(&fc)));
-
-    // Record in live_conns for API visibility.
-    let query = Arc::new(FlowConnectionQuery::new(
-        source_name.to_string(),
-        relationship.to_string(),
-        dest_name.to_string(),
-        Arc::clone(&fc),
-    ));
-    live_conns.write().push(ConnectionInfo {
-        id: conn_id.clone(),
-        source_name: source_name.to_string(),
-        relationship: relationship.to_string(),
-        dest_name: dest_name.to_string(),
-        connection: query,
-    });
-
-    tracing::info!(
-        id = %conn_id,
-        source = source_name,
-        relationship,
-        destination = dest_name,
-        "Hot-added connection"
-    );
-    Ok(conn_id)
-}
-
-fn handle_remove_connection(
-    id: &str,
-    force: bool,
-    live_conns: &Arc<RwLock<Vec<ConnectionInfo>>>,
-) -> std::result::Result<(), MutationError> {
-    let queue_count = {
-        let conns = live_conns.read();
-        let info = conns
-            .iter()
-            .find(|c| c.id == id)
-            .ok_or_else(|| MutationError::ConnectionNotFound(id.to_string()))?;
-        info.connection.queue_count()
-    };
-
-    if queue_count > 0 && !force {
-        return Err(MutationError::QueueNotEmpty(queue_count));
-    }
-
-    if queue_count > 0 {
-        let conns = live_conns.read();
-        if let Some(info) = conns.iter().find(|c| c.id == id) {
-            let removed = info.connection.clear_queue();
-            tracing::warn!(
-                connection_id = %id,
-                discarded = removed,
-                "Force-removing connection with non-empty queue - FlowFiles discarded"
-            );
-        }
-    }
-
-    live_conns.write().retain(|c| c.id != id);
-
-    tracing::info!(id, "Hot-removed connection");
-    Ok(())
 }
