+++ crates/runifi-core/src/engine/mod.rs
@@ -3,5 +3,6 @@ pub mod flow_engine;
 pub mod handle;
 pub mod metrics;
 pub mod mutation;
+pub mod mutation_handler;
 pub mod processor_node;
 pub mod supervisor;
