+++ crates/runifi-core/src/session/process_session.rs
@@ -76,7 +76,6 @@ impl CoreProcessSession {
             yield_duration_ms,
         }
     }
-
 }
 
 impl ProcessSession for CoreProcessSession {
